backend/parallel_api/requirements.txt
# FastAPI and server
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pgvector==0.3.6

# Parallel API client
parallel-sdk==0.4.14

# Azure OpenAI
openai==1.55.3

# Text-to-Speech
elevenlabs==1.10.0

# Environment and utilities
python-dotenv==1.0.1
requests==2.32.3
